pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "visual-language"
version = "0.1.0"
description = "N/A"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [{ name = "ldc-waseda" }]
dependencies = [
  "numpy==2.2.6",
  "pydantic==2.12.0",
  "pillow==11.3.0",
  "openai==1.99.1",
  "matplotlib==3.10.7",
  "aiohttp==3.13.0",
  "opencv-python==4.12.0.88",
  "torch==2.8.0",
  "scikit-learn==1.7.2",
  "Flask==3.1.2",
  "pillow==11.3.0",
  "pandas==2.3.3",
  "seaborn==0.13.2",
  "segment-anything==1.0",
  "sglang==0.5.4",
  "networkx==3.5",
  "openai-batch-wrapper @ git+https://github.com/ldc-waseda/openai-batch-wrapper.git",  # Transparent cache and batch support
]



[project.scripts]
vl-openai = "visual_language.cli.openai_vl:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["visual_language*"]
exclude = ["tests*", "examples*"]
